test product endpoint

+++ repository/repository.go
@@ -7,9 +7,9 @@ import (
 )
 
 type Pagination struct {
-	Offset int `json:"offset"`
-	Limit  int `json:"limit"`
-	Total  int `json:"total"`
+	Offset int   `json:"offset"`
+	Limit  int   `json:"limit"`
+	Total  int64 `json:"total"`
 }
 
 type BulkLoad struct {
